+++ R/swmmr.R
@@ -12,6 +12,9 @@ NULL
 # inspired by https://github.com/jennybc
 # quiets concerns of R CMD check re: the .'s that appear in pipelines
 if (getRversion() >= "2.15.1")  utils::globalVariables(c(".",
+                                                         "Date",
+                                                         "Time",
+                                                         "DateTime",
                                                          "value",
                                                          "Name",
                                                          "pos",